+++ docs/assets/js/fontHelper.js
